Move models into domain
