+++ x-pack/plugins/observability/server/domain/models/duration.test.ts
