+++ x-pack/plugins/observability/server/domain/models/duration.ts
