+++ x-pack/plugins/observability/server/domain/models/error_budget.ts
@@ -4,9 +4,10 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
+
 import * as t from 'io-ts';
-import { dateRangeSchema } from '../schema';
+import { errorBudgetSchema } from '../../types/schema';
 
-type DateRange = t.TypeOf<typeof dateRangeSchema>;
+type ErrorBudget = t.TypeOf<typeof errorBudgetSchema>;
 
-export type { DateRange };
+export type { ErrorBudget };
