+++ x-pack/plugins/observability/server/services/slo/slo_repository.ts
@@ -12,7 +12,7 @@ import { pipe } from 'fp-ts/lib/pipeable';
 import { SavedObjectsClientContract } from '@kbn/core-saved-objects-api-server';
 import { SavedObjectsErrorHelpers } from '@kbn/core-saved-objects-utils-server';
 
-import { StoredSLO, SLO } from '../../types/models';
+import { StoredSLO, SLO } from '../../domain/models';
 import { SO_SLO_TYPE } from '../../saved_objects';
 import { SLONotFound } from '../../errors';
 import { sloSchema } from '../../types/schema';
