+++ x-pack/plugins/observability/server/services/slo/transform_generators/apm_transaction_error_rate.ts
@@ -15,7 +15,7 @@ import {
   SLO_INGEST_PIPELINE_NAME,
   getSLOTransformId,
 } from '../../../assets/constants';
-import { APMTransactionErrorRateIndicator, SLO } from '../../../types/models';
+import { APMTransactionErrorRateIndicator, SLO } from '../../../domain/models';
 
 const APM_SOURCE_INDEX = 'metrics-apm*';
 const ALLOWED_STATUS_CODES = ['2xx', '3xx', '4xx', '5xx'];
