+++ x-pack/plugins/observability/server/services/slo/transform_generators/kql_custom.ts
@@ -17,7 +17,7 @@ import {
   SLO_INGEST_PIPELINE_NAME,
   getSLOTransformId,
 } from '../../../assets/constants';
-import { KQLCustomIndicator, SLO } from '../../../types/models';
+import { KQLCustomIndicator, SLO } from '../../../domain/models';
 
 export class KQLCustomTransformGenerator extends TransformGenerator {
   public getTransformParams(slo: SLO): TransformPutTransformRequest {
