+++ x-pack/plugins/observability/server/types/schema/duration.ts
@@ -7,7 +7,7 @@
 
 import { either } from 'fp-ts/lib/Either';
 import * as t from 'io-ts';
-import { Duration, DurationUnit } from '../models/duration';
+import { Duration, DurationUnit } from '../../domain/models/duration';
 
 const durationType = new t.Type<Duration, string, unknown>(
   'Duration',
